requirements.txt
# 本番依存
playwright==1.44.0
requests==2.31.0
python-dotenv==1.0.1
